+++ src/lib/table/TableSearch.svelte
@@ -25,6 +25,7 @@
     class: className,
     classes,
     placeholder = "Search",
+    oninput,
     ...restProps
   }: TableSearchProps = $props();
 
@@ -57,15 +58,19 @@
   // Handle custom color
   const finalTableClass = $derived(color === "custom" && customColor ? clsx(tableCls, customColor) : tableCls);
 
-  const tableSearchCtx = $derived<TableContextType>({
-    striped,
-    hoverable,
-    color: themeColor
-  });
+  const tableSearchCtx: TableContextType = {
+    get striped() {
+      return striped;
+    },
+    get hoverable() {
+      return hoverable;
+    },
+    get color() {
+      return themeColor;
+    }
+  };
 
-  $effect(() => {
-    setTableContext(tableSearchCtx);
-  });
+  setTableContext(tableSearchCtx);
 </script>
 
 <div class={root({ class: clsx(theme?.root, styling.root) })}>
@@ -77,7 +82,7 @@
           <path fill-rule="evenodd" d="M8 4a4 4 0 100 8 4 4 0 000-8zM2 8a6 6 0 1110.89 3.476l4.817 4.817a1 1 0 01-1.414 1.414l-4.816-4.816A6 6 0 012 8z" clip-rule="evenodd" />
         </svg>
       </div>
-      <input bind:value={inputValue} type="text" id="table-search" class={input({ class: clsx(theme?.input, styling.input) })} {placeholder} />
+      <input bind:value={inputValue} type="text" id="table-search" class={input({ class: clsx(theme?.input, styling.input) })} {placeholder} {oninput} />
     </div>
     {#if header}
       {@render header()}
@@ -117,5 +122,6 @@
 @prop class: className
 @prop classes
 @prop placeholder = "Search"
+@prop oninput
 @prop ...restProps
 -->